Wave function collapse mostly working
It seems like all of the components of the WFC algorithm are working correctly, but the rules I want to set are too strict and lead to impossible tile combinations. The next step will be to implement weights in the rules so that I can place heavy restrictions on what tiles can spawn adjacent, without disabling them completely.