Add an XP and health bar
Player health and experience values represented as UI elements. Health is decremented when the player begins falling past a certain point. XP is awarded for pressing space. After the XP requirement is met, XP is reset to 0 and the requirement increases. These values can be manipulated by different game events with ease.